+++ packages/veritone-csp-generator/CHANGELOG.md
@@ -3,3 +3,7 @@
 ## 3.3.3
 
 * First official release. Changelog, readme, first test added.
+
+## 3.5.7
+
+* update lodash